% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{HSTreeClassifier}
\alias{HSTreeClassifier}
\title{HSTree fit for classification}
\usage{
HSTreeClassifier(X, y, reg_param=1, max_leaf_nodes=20, interaction.depth=1, estimator="CART", shrinkage="node_based")
}
\arguments{
\item{X}{the design matrix}

\item{y}{the response vector}

\item{reg_param}{Higher is more regularization (can be arbitrarily large, should not be < 0)}

\item{max_leaf_nodes}{the maximum number of leaf nodes, the default is 20}

\item{interaction.depth}{Integer specifying the maximum depth of each tree (i.e., the highest level of variable interactions allowed). A value of 1 implies an additive model, a value of 2 implies a model with up to 2-way interactions, etc. Default is 1.}

\item{estimator}{CART decision tree or tree ensemble model (e.g. RandomForest or GradientBoosting)
Defaults to CART Classification Tree with 20 max leaf nodes
Note: this estimator will be directly modified and keep its original functions}

\item{shrinkage}{shrinkage methods, default is "node_based", options are:
1. node_based shrinks based on number of samples in parent node
2. leaf_based only shrinks leaf nodes based on number of leaf samples
3. constant shrinks every node by a constant lambda}
}
\value{
call: the input setting

fit: keep all the output as the orginal estimator, only replace the value in each node based on shrinkage methods

regularization: chr "HSTree"

shrinkage: the hierarchical shrinkage method used in this model

class: the estimator class
}
\description{
modify the decision tree(or each tree in ensemble) structure based on hierarchical shrinkage regularization
}
